fix: compilation errors

+++ src/uniswap-v3/lib/OracleLibrary.sol
@@ -2,8 +2,10 @@
 pragma solidity >=0.5.0;
 
 import {FullMath} from "timeless/lib/FullMath.sol";
-import {TickMath} from "v3-core/libraries/TickMath.sol";
+
 import {IUniswapV3Pool} from "v3-core/interfaces/IUniswapV3Pool.sol";
+
+import {TickMath} from "./TickMath.sol";
 
 /// @title Oracle library
 /// @notice Provides functions to integrate with V3 pool oracle
@@ -34,10 +36,13 @@ library OracleLibrary {
                 1
             ] - secondsPerLiquidityCumulativeX128s[0];
 
-        arithmeticMeanTick = int24(tickCumulativesDelta / secondsAgo);
+        arithmeticMeanTick = int24(
+            tickCumulativesDelta / int56(uint56(secondsAgo))
+        );
         // Always round to negative infinity
         if (
-            tickCumulativesDelta < 0 && (tickCumulativesDelta % secondsAgo != 0)
+            tickCumulativesDelta < 0 &&
+            (tickCumulativesDelta % int56(uint56(secondsAgo)) != 0)
         ) arithmeticMeanTick--;
 
         // We are multiplying here instead of shifting to ensure that harmonicMeanLiquidity doesn't overflow uint128
@@ -159,7 +164,9 @@ library OracleLibrary {
         require(prevInitialized, "ONI");
 
         uint32 delta = observationTimestamp - prevObservationTimestamp;
-        tick = int24((tickCumulative - prevTickCumulative) / delta);
+        tick = int24(
+            (tickCumulative - prevTickCumulative) / int56(uint56(delta))
+        );
         uint128 liquidity = uint128(
             (uint192(delta) * type(uint160).max) /
                 (uint192(
@@ -195,7 +202,7 @@ library OracleLibrary {
         for (uint256 i; i < weightedTickData.length; i++) {
             numerator +=
                 weightedTickData[i].tick *
-                int256(weightedTickData[i].weight);
+                int256(int128(weightedTickData[i].weight));
             denominator += weightedTickData[i].weight;
         }
